requirements.txt
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.4.0
langchain-text-splitters>=1.0.0
langchain-classic>=1.0.0
faiss-cpu>=1.7.4
openai>=1.10.0
python-dotenv>=1.0.0
elevenlabs>=2.0.0
deepgram-sdk>=5.0.0
streamlit>=1.31.0
audio-recorder-streamlit>=0.0.10
